requirements.txt
asgiref==3.4.1
backcall==0.2.0
certifi==2021.10.8
charset-normalizer==2.0.12
colorama==0.4.4
debugpy==1.3.0
decorator==5.0.9
Django==4.0.3
django-crispy-forms==1.14.0
django-session-timeout==0.1.0
djangorestframework==3.13.1
idna==3.3
ipykernel==6.0.1
ipython==7.25.0
ipython-genutils==0.2.0
jedi==0.18.0
jupyter-client==6.1.12
jupyter-core==4.7.1
matplotlib-inline==0.1.2
mysqlclient==2.1.0
numpy==1.21.0
opencv-python==4.5.3.56
parso==0.8.2
pickleshare==0.7.5
prompt-toolkit==3.0.19
Pygments==2.9.0
pyst2==0.5.1
python-dateutil==2.8.1
python-decouple==3.6
pytz==2021.1
pywin32==301
pyzmq==22.1.0
requests==2.27.1
six==1.16.0
sqlparse==0.4.1
tornado==6.1
traitlets==5.0.5
tzdata==2021.5
urllib3==1.26.8
wcwidth==0.2.5
psutil==5.9.0